pyproject.toml
[tool.poetry]
name = "crop_images"
version = "0.1.0"
description = ""
authors = ["EveIsSim"]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^8.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
